% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yenpathy-package.R
\docType{package}
\name{yenpathy}
\alias{yenpathy}
\alias{yenpathy-package}
\title{yenpathy: Quickly Find K Shortest Paths Through a Weighted Graph}
\description{
yenpathy lets you quickly find k shortest paths through even large weighted
graphs. See GitHub for more details at
\url{https://github.com/ecohealthalliance/yenpathy}.
}
\keyword{internal}
